backup_20260110_140811/requirements.extras.txt
# database/auth/data tooling (NOT required for Phase 1 engine + API)
prisma==0.11.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
python-dateutil==2.8.2
aiohttp==3.9.1

# pandas currently breaks on Python 3.14 for this pinned version
# pandas==2.1.4
